+++ drift/lib/src/web/broadcast_stream_queries.dart
@@ -1,34 +1,47 @@
+@JS()
+library;
+
 import 'dart:async';
-import 'dart:html';
+import 'dart:js_interop';
+import 'dart:js_interop_unsafe';
 
 import 'package:drift/src/runtime/api/runtime_api.dart';
 import 'package:drift/src/runtime/executor/stream_queries.dart';
-import 'package:js/js.dart';
-import 'package:js/js_util.dart';
+import 'package:web/web.dart' as web;
+
+@JS('Array')
+extension type _ArrayWrapper._(JSArray _) implements JSObject {
+  external static JSBoolean isArray(JSAny? value);
+}
 
 /// A [StreamQueryStore] using [web broadcast] APIs
 ///
 /// [web broadcast]: https://developer.mozilla.org/en-US/docs/Web/API/Broadcast_Channel_API
 class BroadcastStreamQueryStore extends StreamQueryStore {
-  final BroadcastChannel _channel;
-  StreamSubscription<MessageEvent>? _messageFromChannel;
+  final web.BroadcastChannel _channel;
+  StreamSubscription<web.MessageEvent>? _messageFromChannel;
 
   /// Constructs a broadcast query store with the given [identifier].
   ///
   /// All query stores with the same identifier will share stream query updates.
   BroadcastStreamQueryStore(String identifier)
-      : _channel = BroadcastChannel('drift_updates_$identifier') {
-    _messageFromChannel = _channel.onMessage.listen(_handleMessage);
+      : _channel = web.BroadcastChannel('drift_updates_$identifier') {
+    _messageFromChannel = web.EventStreamProviders.messageEvent
+        .forTarget(_channel)
+        .listen(_handleMessage);
   }
 
-  void _handleMessage(MessageEvent message) {
-    // Using getProperty to avoid dart2js structured clone that turns the
-    // anonymous object into a map.
-    final data = getProperty<Object?>(message, 'data');
-    if (data is! List || data.isEmpty) return;
+  void _handleMessage(web.MessageEvent message) {
+    final data = message.data;
+    if (!_ArrayWrapper.isArray(data).toDart) {
+      return;
+    }
+
+    final asList = (data as JSArray).toDart;
+    if (asList.isEmpty) return;
 
     super.handleTableUpdates({
-      for (final entry in data.cast<_SerializedTableUpdate>())
+      for (final entry in asList.cast<_SerializedTableUpdate>())
         entry.toTableUpdate,
     });
   }
@@ -39,7 +52,7 @@ class BroadcastStreamQueryStore extends StreamQueryStore {
 
     _channel.postMessage([
       for (final update in updates) _SerializedTableUpdate.of(update),
-    ]);
+    ].toJS);
   }
 
   @override
@@ -50,34 +63,31 @@ class BroadcastStreamQueryStore extends StreamQueryStore {
   }
 
   /// Whether the current JavaScript context supports broadcast channels.
-  static bool get supported => hasProperty(globalThis, 'BroadcastChannel');
+  static bool get supported => globalContext.has('BroadcastChannel');
 }
 
 @JS()
 @anonymous
-@staticInterop
-class _SerializedTableUpdate {
+extension type _SerializedTableUpdate._(JSObject _) implements JSObject {
   external factory _SerializedTableUpdate({
-    required String? kind,
-    required String table,
+    required JSString? kind,
+    required JSString table,
   });
 
   factory _SerializedTableUpdate.of(TableUpdate update) {
-    return _SerializedTableUpdate(kind: update.kind?.name, table: update.table);
+    return _SerializedTableUpdate(
+      kind: update.kind?.name.toJS,
+      table: update.table.toJS,
+    );
   }
-}
 
-extension on _SerializedTableUpdate {
-  @JS()
-  external String? get kind;
-
-  @JS()
-  external String get table;
+  external JSString? get kind;
+  external JSString get table;
 
   TableUpdate get toTableUpdate {
-    final updateKind = _updateKindByName[kind];
+    final updateKind = _updateKindByName[kind?.toDart];
 
-    return TableUpdate(table, kind: updateKind);
+    return TableUpdate(table.toDart, kind: updateKind);
   }
 
   static final _updateKindByName = UpdateKind.values.asNameMap();