+++ drift/lib/src/web/wasm_setup/shared.dart
@@ -1,17 +1,25 @@
 import 'dart:async';
-import 'dart:html';
-import 'dart:indexed_db';
+import 'dart:js_interop';
+import 'dart:js_interop_unsafe';
 
 import 'package:drift/drift.dart';
 import 'package:drift/remote.dart';
 import 'package:drift/wasm.dart';
-import 'package:js/js_util.dart';
+import 'package:web/web.dart'
+    show
+        Worker,
+        IDBFactory,
+        IDBRequest,
+        IDBDatabase,
+        IDBVersionChangeEvent,
+        EventStreamProviders,
+        MessageEvent;
 // ignore: implementation_imports
 import 'package:sqlite3/src/wasm/js_interop/file_system_access.dart';
 import 'package:sqlite3/wasm.dart';
 import 'package:stream_channel/stream_channel.dart';
 
-import '../channel.dart';
+import '../new_channel.dart';
 import 'protocol.dart';
 
 /// Checks whether the OPFS API is likely to be correctly implemented in the
@@ -38,10 +46,10 @@ Future<bool> checkOpfsSupport() async {
     // In earlier versions of the OPFS standard, some methods like `getSize()`
     // on a sync file handle have actually been asynchronous. We don't support
     // Browsers that implement the outdated spec.
-    final getSizeResult = callMethod<Object?>(openedFile, 'getSize', []);
-    if (typeofEquals<Object?>(getSizeResult, 'object')) {
+    final getSizeResult = (openedFile as JSObject).callMethod('getSize'.toJS);
+    if (getSizeResult.typeofEquals('object')) {
       // Returned a promise, that's no good.
-      await promiseToFuture<Object?>(getSizeResult!);
+      await (getSizeResult as JSPromise).toDart;
       return false;
     }
 
@@ -61,18 +69,18 @@ Future<bool> checkOpfsSupport() async {
 
 /// Checks whether IndexedDB is working in the current browser.
 Future<bool> checkIndexedDbSupport() async {
-  if (!hasProperty(globalThis, 'indexedDB') ||
+  if (!globalContext.has('indexedDB') ||
       // FileReader needed to read and write blobs efficiently
-      !hasProperty(globalThis, 'FileReader')) {
+      !globalContext.has('FileReader')) {
     return false;
   }
 
-  final idb = getProperty<IdbFactory>(globalThis, 'indexedDB');
+  final idb = globalContext['indexedDB'] as IDBFactory;
 
   try {
     const name = 'drift_mock_db';
 
-    final mockDb = await idb.open(name);
+    final mockDb = await idb.open(name).complete<IDBDatabase>();
     mockDb.close();
     idb.deleteDatabase(name);
   } catch (error) {
@@ -87,19 +95,16 @@ Future<bool> checkIndexedDbExists(String databaseName) async {
   bool? indexedDbExists;
 
   try {
-    final idb = getProperty<IdbFactory>(globalThis, 'indexedDB');
+    final idb = globalContext['indexedDB'] as IDBFactory;
 
-    final database = await idb.open(
-      databaseName,
-      // Current schema version used by the [IndexedDbFileSystem]
-      version: 1,
-      onUpgradeNeeded: (event) {
-        // If there's an upgrade, we're going from 0 to 1 - the database doesn't
-        // exist! Abort the transaction so that we don't create it here.
-        event.target.transaction!.abort();
-        indexedDbExists = false;
-      },
-    );
+    final openRequest = idb.open(databaseName, 1);
+    openRequest.onupgradeneeded = (IDBVersionChangeEvent event) {
+      // If there's an upgrade, we're going from 0 to 1 - the database doesn't
+      // exist! Abort the transaction so that we don't create it here.
+      openRequest.transaction!.abort();
+      indexedDbExists = false;
+    }.toJS;
+    final database = await openRequest.complete<IDBDatabase>();
 
     indexedDbExists ??= true;
     database.close();
@@ -112,9 +117,9 @@ Future<bool> checkIndexedDbExists(String databaseName) async {
 
 /// Deletes a database from IndexedDb if supported.
 Future<void> deleteDatabaseInIndexedDb(String databaseName) async {
-  final idb = window.indexedDB;
-  if (idb != null) {
-    await idb.deleteDatabase(databaseName);
+  if (globalContext.has('indexedDB')) {
+    final idb = globalContext['indexedDB'] as IDBFactory;
+    await idb.deleteDatabase(databaseName).complete<JSAny?>();
   }
 }
 
@@ -181,12 +186,16 @@ class DriftServerController {
       final initPort = message.initializationPort;
 
       final initializer = initPort != null
-          ? () async {
-              initPort.postMessage(true);
+          ? () {
+              final completer = Completer<Uint8List?>();
+              initPort.postMessage(true.toJS);
 
-              return await initPort.onMessage
-                  .map((e) => e.data as Uint8List?)
-                  .first;
+              initPort.onmessage = (MessageEvent e) {
+                final data = (e.data as JSUint8Array?);
+                completer.complete(data?.toDart);
+              }.toJS;
+
+              return completer.future;
             }
           : null;
 
@@ -269,7 +278,7 @@ class DriftServerController {
     StartFileSystemServer(options).sendToWorker(worker);
 
     // Wait for the server worker to report that it's ready
-    await worker.onMessage.first;
+    await EventStreamProviders.messageEvent.forTarget(worker).first;
 
     return WasmVfs(workerOptions: options);
   }
@@ -348,4 +357,22 @@ extension StorageClassification on WasmStorageImplementation {
   bool get isIndexedDbBased =>
       this == WasmStorageImplementation.sharedIndexedDb ||
       this == WasmStorageImplementation.unsafeIndexedDb;
+}
+
+/// Utilities to complete an IndexedDB request.
+extension CompleteIdbRequest on IDBRequest {
+  /// Turns this request into a Dart future that completes with the first
+  /// success or error event.
+  Future<T> complete<T extends JSAny?>() {
+    final completer = Completer<T>.sync();
+
+    EventStreamProviders.successEvent.forTarget(this).listen((event) {
+      completer.complete(result as T);
+    });
+    EventStreamProviders.errorEvent.forTarget(this).listen((event) {
+      completer.completeError(error ?? event);
+    });
+
+    return completer.future;
+  }
 }